Plot Impulse Response Functions of a VAR

Plots the result of a irf call.

Usage

ggvar_irf(
  x,
  series_imp = NULL,
  series_resp = NULL,
  n.ahead = 10,
  ci = 0.95,
  args_line = list(),
  args_hline = list(),
  args_ribbon = list(fill = NA, linetype = 2, color = "blue"),
  args_facet = list(),
  facet_type = "ggplot",
  ...
)

Arguments

x

A "varest" object to pass to irf, or, directly, a "varirf" object.

series_imp

A character vector with variables to consider for the impulses. Defaults to all (NULL).

series_resp

A character vector with variables to consider for the responses. Defaults to all (NULL).

n.ahead

An integer. The size of the forecast horizon, passed to irf. Unused if x is "varirf".

ci

The level of confidence for the irf. Set to FALSE to omit.

args_line

Additional arguments passed to ggplot2::geom_line.

args_hline

Additional arguments passed to ggplot2::geom_hline.

args_ribbon

Additional arguments passed to ggplot2::geom_ribbon.

args_facet

Additional arguments passed to the faceting engine used.

facet_type

The facet 'engine' to be used. 'ggplot2' for ggplot2::facet_grid, 'ggh4x' for ggh4x::facet_grid2.

...

Arguments passed to varr:::irf_setup, the generic function that formats x into a 'graphable' format. Use them only if you have created a method for some unsupported class of x.Pass additional arguments to vars::irf here.

Value

An object of class 'ggplot'.

Examples

ggvar_irf(vars::VAR(freeny[-2]), args_facet = list(scales = "free_y"))
#> Error in setup_ggvar_irf(x, n.ahead, series_impulse, series_response,     ci, facet): `n.ahead` must be supplied with `x` of class 'varest'